cmx-runtime-environment-dev/cmx-runtime-environment-dev/terraform/traffic-cop-environment/files/elasticsearch_lambda_function/requirements.txt
boto3==1.16.50
botocore==1.19.50
certifi==2020.12.5
chardet==4.0.0
click==6.7
elasticsearch==7.1.0
elasticsearch-curator==5.8.3
futures==3.3.0
idna==2.10
jmespath==0.10.0
pip==20.3.3
python-dateutil==2.8.1
PyYAML==3.13
requests==2.25.1
requests-aws4auth==1.0.1
s3transfer==0.3.3
setuptools==44.1.1
six==1.15.0
urllib3==1.25.11
voluptuous==0.12.1
wheel==0.36.1
